separate metric with numerics and others
