+++ src/main/java/org/elasticsearch/search/reducers/metric/MetricResultFactory.java
@@ -21,17 +21,13 @@
 package org.elasticsearch.search.reducers.metric;
 
 import org.elasticsearch.ElasticsearchParseException;
-import org.elasticsearch.search.reducers.metric.multi.delta.DeltaResult;
-import org.elasticsearch.search.reducers.metric.multi.stats.StatsResult;
-import org.elasticsearch.search.reducers.metric.single.SingleMetricResult;
+import org.elasticsearch.search.reducers.metric.numeric.multi.stats.StatsResult;
+import org.elasticsearch.search.reducers.metric.numeric.single.SingleMetricResult;
 
 public class MetricResultFactory {
     public static MetricResult getInstance(String type) {
         if (type.equals("single_metric")) {
             return new SingleMetricResult();
-        }
-        if (type.equals("delta_metric")) {
-            return new DeltaResult();
         }
         if (type.equals("stats_metric")) {
             return new StatsResult();
